@@ -18,23 +18,48 @@ func registerRoutes(router *fiber.Router) {
 
 	(*router).Get("/health", func(c *fiber.Ctx) error { return c.Status(http.StatusOK).SendString("ok") })
 
-	(*router).Get("/user/me", jwtController.IsConnectedMiddleware(domain.PermissionUser), userController.GetMe)
-	(*router).Get("/user/:uuid", userController.GetName)
+	(*router).Get("/user/me",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), userController.GetMe)
 
-	(*router).Post("/security/login", jwtController.IsConnectedMiddleware(domain.PermissionNone), authController.Login)
-	(*router).Post("/security/register", jwtController.IsConnectedMiddleware(domain.PermissionNone), authController.Register)
-	(*router).Post("/security/logout", jwtController.IsConnectedMiddleware(domain.PermissionUser), authController.Logout)
-	(*router).Post("/security/refresh", jwtController.IsConnectedMiddleware(domain.PermissionUser), authController.RefreshToken)
+	(*router).Get("/user/:uuid",
+		userController.GetName)
 
-	(*router).Get("/security/github", oauthController.GithubLogin)
-	(*router).Get("/security/github_callback", oauthController.GithubCallback)
+	(*router).Post("/security/login",
+		jwtController.IsConnectedMiddleware(domain.PermissionNone), authController.Login)
 
-	(*router).Get("/security/discord", oauthController.DiscordLogin)
-	(*router).Get("/security/discord_callback", oauthController.DiscordCallback)
+	(*router).Post("/security/register",
+		jwtController.IsConnectedMiddleware(domain.PermissionNone), authController.Register)
 
-	(*router).Get("/deck/owned", jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetOwned)
-	(*router).Get("/deck/public", jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetPublic)
-	(*router).Get("/deck/learning", jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetLearning)
-	(*router).Get("/deck/:id", jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetByID)
-	(*router).Post("/deck", jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.Create)
+	(*router).Post("/security/logout",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), authController.Logout)
+
+	(*router).Post("/security/refresh",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), authController.RefreshToken)
+
+	(*router).Get("/security/github",
+		oauthController.GithubLogin)
+
+	(*router).Get("/security/github_callback",
+		oauthController.GithubCallback)
+
+	(*router).Get("/security/discord",
+		oauthController.DiscordLogin)
+
+	(*router).Get("/security/discord_callback",
+		oauthController.DiscordCallback)
+
+	(*router).Get("/deck/owned",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetOwned)
+
+	(*router).Get("/deck/public",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetPublic)
+
+	(*router).Get("/deck/learning",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetLearning)
+
+	(*router).Get("/deck/:id",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.GetByID)
+
+	(*router).Post("/deck",
+		jwtController.IsConnectedMiddleware(domain.PermissionUser), deckController.Create)
 }
