@@ -4,7 +4,7 @@ import (
 	"gorm.io/gorm"
 )
 
-// User is the domain model for a user
+// User is the domain model for a user.
 type User struct {
 	gorm.Model    `swaggerignore:"true"`
 	Username      string     `json:"username"`
@@ -19,12 +19,12 @@ type User struct {
 	Oauth         bool       `json:"oauth" gorm:"default:false"`
 }
 
-// TableName returns the table name for the user model
+// TableName returns the table name for the user model.
 func (*User) TableName() string {
 	return "users"
 }
 
-// ToPublicUser converts the user to a public user
+// ToPublicUser converts the user to a public user.
 func (u *User) ToPublicUser() PublicUser {
 	return PublicUser{
 		ID:         u.ID,
@@ -35,17 +35,17 @@ func (u *User) ToPublicUser() PublicUser {
 	}
 }
 
-// Validate validates the user
+// Validate validates the user.
 func (u *User) Validate() error {
-	return validate.Struct(u)
+	return GetValidatorInstance().Validate().Struct(u)
 }
 
-// HasPermission checks if the user has the given permission
+// HasPermission checks if the user has the given permission.
 func (u *User) HasPermission(permission Permission) bool {
 	return u.Permission >= permission
 }
 
-// PublicUser is the public user model
+// PublicUser is the public user model.
 type PublicUser struct {
 	Username   string     `json:"username"`   // Username of the user
 	Email      string     `json:"email"`      // Email of the user
@@ -54,25 +54,25 @@ type PublicUser struct {
 	Permission Permission `json:"permission"` // Permission of the user
 }
 
-// Login is the login model
+// Login is the login model.
 type Login struct {
 	Email    string `json:"email" validate:"email"` // Email of the user
 	Password string `json:"password"`               // Password of the user
 }
 
-// Register is the register model
+// Register is the register model.
 type Register struct {
 	Username string `json:"username" validate:"required"` // Username of the user
 	Email    string `json:"email" validate:"email"`       // Email of the user
 	Password string `json:"password" validate:"required"` // Password of the user
 }
 
-// Validate validates the register model
+// Validate validates the register model.
 func (r *Register) Validate() error {
-	return validate.Struct(r)
+	return GetValidatorInstance().Validate().Struct(r)
 }
 
-// ToUser converts the register model to a user
+// ToUser converts the register model to a user.
 func (r *Register) ToUser() User {
 	return User{
 		Username:   r.Username,
