@@ -2,18 +2,29 @@ package infrastructures
 
 import (
 	"context"
+	"log/slog"
+	"sync"
 
 	"github.com/getsentry/sentry-go"
 	sentryotel "github.com/getsentry/sentry-go/otel"
+	"github.com/gofiber/fiber/v2/log"
 	"github.com/memnix/memnix-rest/config"
-	"github.com/uptrace/opentelemetry-go-extra/otelzap"
 	"go.opentelemetry.io/otel"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/trace"
-	"go.uber.org/zap"
 )
 
-var fiberTracer = otel.Tracer("fiber-server")
+var (
+	once     sync.Once    //nolint:gochecknoglobals // Singleton
+	instance trace.Tracer //nolint:gochecknoglobals // Singleton
+)
+
+func GetTracerInstance() trace.Tracer {
+	once.Do(func() {
+		instance = otel.Tracer("fiber-server")
+	})
+	return instance
+}
 
 func InitTracer(cfg config.SentryConfig) error {
 	initSentry(cfg)
@@ -31,12 +42,10 @@ func ShutdownTracer() error {
 	return nil
 }
 
-func GetFiberTracer() trace.Tracer {
-	return fiberTracer
-}
-
 func initSentry(cfg config.SentryConfig) {
-	otelzap.Ctx(context.Background()).Info("Initializing Sentry :", zap.Float64("traces_sample_rate", cfg.TracesSampleRate), zap.Float64("profiles_sample_rate", cfg.ProfilesSampleRate))
+	log.WithContext(context.Background()).Info(
+		"Initializing Sentry :", slog.Float64("traces_sample_rate", cfg.TracesSampleRate), slog.Float64(
+			"profiles_sample_rate", cfg.ProfilesSampleRate))
 	_ = sentry.Init(sentry.ClientOptions{
 		Dsn:                cfg.DSN,
 		Debug:              cfg.Debug,
