@@ -5,19 +5,19 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"log/slog"
 	"net/http"
 
+	"github.com/gofiber/fiber/v2/log"
 	"github.com/pkg/errors"
-	"github.com/uptrace/opentelemetry-go-extra/otelzap"
-	"go.uber.org/zap"
 )
 
 const (
-	discordTokenURL = "https://discord.com/api/oauth2/token"
-	discordAPIURL   = "https://discord.com/api/users/@me"
+	discordTokenURL = "https://discord.com/api/oauth2/token" //nolint:gosec //This is a URL, not a password.
+	discordAPIURL   = "https://discord.com/api/users/@me"    //nolint:gosec //This is a URL, not a password.
 )
 
-// Represents the response received from Discord
+// Represents the response received from Discord.
 type discordAccessTokenResponse struct {
 	AccessToken  string `json:"access_token"`
 	TokenType    string `json:"token_type"`
@@ -26,39 +26,39 @@ type discordAccessTokenResponse struct {
 	RefreshToken string `json:"refresh_token"`
 }
 
-// GetDiscordAccessToken gets the access token from Discord
+// GetDiscordAccessToken gets the access token from Discord.
 func GetDiscordAccessToken(ctx context.Context, code string) (string, error) {
-	reqBody := bytes.NewBuffer([]byte(fmt.Sprintf(
+	reqBody := bytes.NewBufferString(fmt.Sprintf(
 		"client_id=%s&client_secret=%s&grant_type=authorization_code&redirect_uri=%s&code=%s&scope=identify,email",
 		discordConfig.ClientID,
 		discordConfig.ClientSecret,
 		GetCallbackURL()+"/v2/security/discord_callback",
 		code,
-	)))
+	))
 
-	// POST request to set URL
+	// POST request to set URL.
 	req, err := http.NewRequestWithContext(ctx,
 		http.MethodPost,
 		discordTokenURL,
 		reqBody,
 	)
 	if err != nil {
-		otelzap.Ctx(ctx).Error("Failed to get Discord access token", zap.Error(err))
+		log.WithContext(ctx).Error("Failed to get Discord access token", slog.Any("error", err))
 		return "", errors.Wrap(err, RequestFailed)
 	}
 
 	if req == nil || req.Body == nil || req.Header == nil {
-		otelzap.Ctx(ctx).Error("Failed to get Discord access token", zap.Error(err))
+		log.WithContext(ctx).Error("Failed to get Discord access token", slog.Any("error", err))
 		return "", errors.New(RequestFailed)
 	}
 
 	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	req.Header.Set("Accept", "application/json")
 
-	// Get the response
+	// Get the response.
 	resp, resperr := http.DefaultClient.Do(req)
 	if resperr != nil {
-		otelzap.Ctx(ctx).Error("Failed to get Discord access token", zap.Error(resperr))
+		log.WithContext(ctx).Error("Failed to get Discord access token", slog.Any("error", resperr))
 		return "", errors.Wrap(resperr, ResponseFailed)
 	}
 
@@ -68,26 +68,26 @@ func GetDiscordAccessToken(ctx context.Context, code string) (string, error) {
 		}
 	}(resp)
 
-	// Response body converted to stringified JSON
+	// Response body converted to stringified JSON.
 	respbody, err := io.ReadAll(resp.Body)
 	if err != nil {
-		otelzap.Ctx(ctx).Error("failed to read resp.body", zap.Error(err))
+		log.WithContext(ctx).Error("failed to read resp.body", slog.Any("error", err))
 		return "", err
 	}
 
-	// Convert stringified JSON to a struct object of type githubAccessTokenResponse
+	// Convert stringified JSON to a struct object of type githubAccessTokenResponse.
 	var ghresp discordAccessTokenResponse
-	err = jsonHelper.Unmarshal(respbody, &ghresp)
+	err = GetJSONHelperInstance().GetJSONHelper().Unmarshal(respbody, &ghresp)
 	if err != nil {
 		return "", err
 	}
 
 	// Return the access token (as the rest of the
-	// details are relatively unnecessary for us)
+	// details are relatively unnecessary for us).
 	return ghresp.AccessToken, nil
 }
 
-// GetDiscordData gets the user data from Discord
+// GetDiscordData gets the user data from Discord.
 func GetDiscordData(ctx context.Context, accessToken string) (string, error) {
 	req, err := http.NewRequestWithContext(ctx,
 		http.MethodGet,
@@ -100,7 +100,7 @@ func GetDiscordData(ctx context.Context, accessToken string) (string, error) {
 
 	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", accessToken))
 
-	// Get the response
+	// Get the response.
 	resp, err := http.DefaultClient.Do(req)
 	if err != nil {
 		return "", errors.Wrap(err, ResponseFailed)
@@ -111,7 +111,7 @@ func GetDiscordData(ctx context.Context, accessToken string) (string, error) {
 			resp.Body.Close()
 		}
 	}(resp)
-	// Response body converted to stringified JSON
+	// Response body converted to stringified JSON.
 	respbody, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return "", err
