@@ -11,18 +11,18 @@ import (
 )
 
 const (
-	githubAccessTokenURL = "https://github.com/login/oauth/access_token"
-	githubAPIURL         = "https://api.github.com/user"
+	githubAccessTokenURL = "https://github.com/login/oauth/access_token" //nolint:gosec //This is a URL, not a password.
+	githubAPIURL         = "https://api.github.com/user"                 //nolint:gosec //This is a URL, not a password.
 )
 
-// githubAccessTokenResponse Represents the response received from Github
+// githubAccessTokenResponse Represents the response received from Github.
 type githubAccessTokenResponse struct {
 	AccessToken string `json:"access_token"`
 	TokenType   string `json:"token_type"`
 	Scope       string `json:"scope"`
 }
 
-// GetGithubAccessToken gets the access token from Github using the code
+// GetGithubAccessToken gets the access token from Github using the code.
 func GetGithubAccessToken(ctx context.Context, code string) (string, error) {
 	// Set us the request body as JSON
 	requestBodyMap := map[string]string{
@@ -30,7 +30,7 @@ func GetGithubAccessToken(ctx context.Context, code string) (string, error) {
 		"client_secret": githubConfig.ClientSecret,
 		"code":          code,
 	}
-	requestJSON, _ := jsonHelper.Marshal(requestBodyMap)
+	requestJSON, _ := GetJSONHelperInstance().GetJSONHelper().Marshal(requestBodyMap)
 
 	// POST request to set URL
 	req, err := http.NewRequestWithContext(ctx,
@@ -55,12 +55,12 @@ func GetGithubAccessToken(ctx context.Context, code string) (string, error) {
 			resp.Body.Close()
 		}
 	}(resp)
-	// Response body converted to stringified JSON
+	// Response body converted to stringified JSON.
 	respbody, _ := io.ReadAll(resp.Body)
 
 	// Convert stringified JSON to a struct object of type githubAccessTokenResponse
 	var ghresp githubAccessTokenResponse
-	err = jsonHelper.Unmarshal(respbody, &ghresp)
+	err = GetJSONHelperInstance().GetJSONHelper().Unmarshal(respbody, &ghresp)
 	if err != nil {
 		return "", err
 	}
@@ -70,7 +70,7 @@ func GetGithubAccessToken(ctx context.Context, code string) (string, error) {
 	return ghresp.AccessToken, nil
 }
 
-// GetGithubData gets the user data from Github using the access token
+// GetGithubData gets the user data from Github using the access token.
 func GetGithubData(ctx context.Context, accessToken string) (string, error) {
 	// Get request to a set URL
 	req, err := http.NewRequestWithContext(ctx,
