@@ -9,7 +9,7 @@ const (
 	bitSize = 32
 )
 
-// ConvertStrToUInt converts a string to an unsigned integer
+// ConvertStrToUInt converts a string to an unsigned integer.
 func ConvertStrToUInt(str string) (uint, error) {
 	number, err := strconv.ParseUint(str, base10, bitSize)
 	if err != nil {
@@ -18,12 +18,12 @@ func ConvertStrToUInt(str string) (uint, error) {
 	return uint(number), nil
 }
 
-// ConvertUIntToStr converts an unsigned integer to a string
+// ConvertUIntToStr converts an unsigned integer to a string.
 func ConvertUIntToStr(number uint) string {
 	return strconv.FormatUint(uint64(number), base10)
 }
 
-// ConvertStrToInt converts a string to an integer
+// ConvertStrToInt converts a string to an integer.
 func ConvertStrToInt(str string) (int, error) {
 	number, err := strconv.ParseInt(str, base10, bitSize)
 	if err != nil {
