@@ -2,7 +2,7 @@ package views
 
 const (
 	// InvalidCredentials is the message for invalid credentials.
-	InvalidCredentials = "invalid credentials" //nolint:gosec
+	InvalidCredentials = "invalid credentials" //nolint:gosec //This is a message, not a password.
 	RequestFailed      = "request failed"
 	ResponseFailed     = "response failed"
 )
